+++ test/stirrer.tests.js
@@ -1,5 +1,7 @@
 /* jshint strict: false */
 
+require("../lib");  //sets the base for coverage - require all the codes
+
 require("./utils.test");